# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8000
info:
  contact:
    email: support@xilo.example.com
    name: Xilo API Support
    url: https://xilo.example.com/support
  description: A self-hosted, Telegram-inspired modern blog platform API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://xilo.example.com/terms
  title: Xilo Blog Platform API
  version: "1.0"
paths:
...
  /billing/callback:
    get:
      parameters:
      - description: Payment authority
        in: query
        name: Authority
        required: true
        type: string
      - description: Payment status
        in: query
        name: Status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect
          schema:
            type: string
      summary: Payment gateway callback
      tags:
      - billing
  /billing/invoices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user invoices
      tags:
      - billing
  /billing/my-subscription:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user subscription
      tags:
      - billing
  /billing/plans:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List subscription plans
      tags:
      - billing
    post:
      consumes:
      - application/json
      parameters:
      - description: Plan data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a subscription plan
      tags:
      - billing
  /billing/plans/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated plan data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a subscription plan
      tags:
      - billing
  /billing/plans/{slug}:
    get:
      parameters:
      - description: Plan slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get plan by slug
      tags:
      - billing
...
  /billing/subscribe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe to a plan
      tags:
      - billing
  /billing/subscription:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel current subscription
      tags:
      - billing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"